AUV_VIEWER new harvester
This harvester intends to replace the old process to populate the auv viewer,
as well as the current legacy_auv schema in dbprod.
Part of this harvester is the matlab process (see in data-services repo)
The harvester reads the output files created by the matlab script found in
/mnt/ebs/wip/AUV/auv_viewer_process
The files in this folder are currently symbolic links since the matlab script
has run for more than 10 days in a row. All files will be moved physically
there in a near future, but the harvester can at least be tested and reviewed

Thanks, loz