AUV VIEWER GEOM CONSTRAINTS
removed CHECK (st_isvalid(geom)) on both auv_images and tracks tables.
If there was one bad image geom in one dive (approx 6000 images), then
a die on error would not process the dive. Instead there is a makevalid
(geom) on both views.
ready to be merged and tested on RC